remove DocumentInterface, resolve #36, resolve #30
