+++ phpstan-baseline.neon
@@ -36,19 +36,9 @@ parameters:
 			path: src/Command/Refresh.php
 
 		-
-			message: "#^Method Valantic\\\\ElasticaBridgeBundle\\\\DocumentType\\\\AbstractDocument\\:\\:getListingClass\\(\\) should return class\\-string but returns string\\.$#"
+			message: "#^Method Valantic\\\\ElasticaBridgeBundle\\\\DocumentType\\\\Index\\\\AbstractIndexDocument\\:\\:getListingClass\\(\\) should return class\\-string but returns string\\.$#"
 			count: 1
-			path: src/DocumentType/AbstractDocument.php
-
-		-
-			message: "#^Method Valantic\\\\ElasticaBridgeBundle\\\\Repository\\\\DocumentRepository\\:\\:__construct\\(\\) has parameter \\$documents with no value type specified in iterable type iterable\\.$#"
-			count: 1
-			path: src/Repository/DocumentRepository.php
-
-		-
-			message: "#^Property Valantic\\\\ElasticaBridgeBundle\\\\Repository\\\\DocumentRepository\\:\\:\\$documents \\(array\\<Valantic\\\\ElasticaBridgeBundle\\\\DocumentType\\\\DocumentInterface\\>\\) does not accept array\\<string, object\\>\\.$#"
-			count: 1
-			path: src/Repository/DocumentRepository.php
+			path: src/DocumentType/Index/AbstractIndexDocument.php
 
 		-
 			message: "#^Method Valantic\\\\ElasticaBridgeBundle\\\\Repository\\\\IndexDocumentRepository\\:\\:__construct\\(\\) has parameter \\$indexDocuments with no value type specified in iterable type iterable\\.$#"
