+++ src/DependencyInjection/ValanticElasticaBridgeExtension.php
@@ -8,7 +8,6 @@ use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-use Valantic\ElasticaBridgeBundle\DocumentType\DocumentInterface;
 use Valantic\ElasticaBridgeBundle\DocumentType\Index\IndexDocumentInterface;
 use Valantic\ElasticaBridgeBundle\Index\IndexInterface;
 
@@ -20,7 +19,6 @@ use Valantic\ElasticaBridgeBundle\Index\IndexInterface;
 class ValanticElasticaBridgeExtension extends Extension
 {
     private const TAG_INDEX = 'valantic.elastica_bridge.index';
-    private const TAG_DOCUMENT = 'valantic.elastica_bridge.document';
     private const TAG_DOCUMENT_INDEX = 'valantic.elastica_bridge.document_index';
 
     /**
@@ -33,7 +31,6 @@ class ValanticElasticaBridgeExtension extends Extension
     public function load(array $configs, ContainerBuilder $container): void
     {
         $container->registerForAutoconfiguration(IndexInterface::class)->addTag(self::TAG_INDEX);
-        $container->registerForAutoconfiguration(DocumentInterface::class)->addTag(self::TAG_DOCUMENT);
         $container->registerForAutoconfiguration(IndexDocumentInterface::class)->addTag(self::TAG_DOCUMENT_INDEX);
 
         $configuration = new Configuration();
