+++ src/Resources/config/services.yml
@@ -45,9 +45,6 @@ services:
     arguments:
       - !tagged_iterator valantic.elastica_bridge.index
 
-  Valantic\ElasticaBridgeBundle\Repository\DocumentRepository:
-    arguments:
-      - !tagged_iterator valantic.elastica_bridge.document
 
   Valantic\ElasticaBridgeBundle\Repository\IndexDocumentRepository:
     arguments:
